// core-libs/core/src/model/product.model.ts
export interface Review {
  alias?: string;
  comment?: string;
  date?: Date;
  headline?: string;
  id?: string;
  principal?: User;
  rating?: number;
}